4.21 Step 9: Delete with token authorization

+++ part4/bloglist-backend/controllers/blogs.js
@@ -34,8 +34,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) return response.status(401).json({ error: 'blog not found' })
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'missing or invalid token' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
